==> 00 - Datasets Creation and Analysis.ipynb <==
import json 
import numpy as np
import pandas as pd

#!pip install fasttext
#!wget -O ./data/lid.176.bin https://dl.fbaipublicfiles.com/fasttext/supervised-models/lid.176.bin
import fasttext
import re

import string

from nltk.corpus import stopwords
from nltk.stem import SnowballStemmer
from nltk.stem import WordNetLemmatizer
from nltk.tokenize import word_tokenize

## [WH+TW] Create datasets

def get_whisper(filename):
    serial = []
    text = []
    for line in open(filename, 'r'):
        new_data = json.loads(line)
        serial.append(new_data["serial"])
        text.append(new_data["text"])
    return pd.DataFrame({"serial": serial, "text": text})

def get_twitter(filename):
    text = []
    for line in open(filename, 'r'):
        splitted = line.split('\t')
        if len(splitted) == 4:
            text.append(splitted[2])
    return pd.DataFrame({"text": text})

def is_english(text):
    dl = lang_model.predict(text)[0][0][-2:]
    return dl == 'en'

def replace_if_lang(df, text):
    def func(row):
        if is_english(row["text"]):
            return row["text"]
        return ""
    df['text'] = df.apply(func, axis=1).tolist()
    return df

def replace_in_df(df, regex, text):
    def func(row):
        return re.sub(regex, text, row["text"])
    df['text'] = df.apply(func, axis=1).tolist()
    return df

def clean_tweets(df, min_length=8):
    # delete the entire tweet (1/2)
    df = replace_in_df(df, r".*RT @.*", "") # retweets
    df = replace_in_df(df, r".*\[\[.*", "") # placeholders
    # delete only a part of the tweet
    df = replace_in_df(df, r"http(s?)://\S+", "") # links
    df = replace_in_df(df, r"@\S+", "") # mentions
    df = replace_in_df(df, r"#\S+", "") # hashtags
    df = replace_in_df(df, r"\n", " ") # newline
    df = replace_in_df(df, r"  +", " ") # multiple spaces
    df = replace_in_df(df, r"^ ", "") # space at the begin of line
    df = replace_in_df(df, r" $", "") # space at the end of line
    # delete the entire tweet (2/2)
    df = replace_if_lang(df, "") # texts not in English
    df = df[df['text'].map(len) >= min_length] # short text
    return df.drop_duplicates()

def get_datasets(whisper_filename, twitter_filename):
    def get_dataset(filename, func, cls):
        data = func(filename)
        data["class"] = cls
        data = data[['text', 'class']]
        return clean_tweets(data)
    data_w = get_dataset(whisper_filename, get_whisper, "sens")
    data_t = get_dataset(twitter_filename, get_twitter, "ns")
    return data_w, data_t

def save_samples_data(num_samples, df_sens, df_ns, num_sens, num_ns, filename_prefix):
    for count in range(num_samples):
        df_sens[(num_sens*count):(num_sens*(count+1))].append(df_ns[(num_ns*count):(num_ns*(count+1))]).sample(frac=1, random_state=2**count).reset_index(drop=True).to_csv(filename_prefix+("0"+str(count+1))[-2:]+".csv", index=False)
    return

lang_model = fasttext.load_model('./data/lid.176.bin')


df_sens, df_ns = get_datasets('/data/whisper/final-anonymized.jun14.jun16.whisper-part-000.json', 
                '/data/twitter-cikm-2010/test_set_tweets.txt')

df_sens

df_ns

save_samples_data(10, df_sens.sample(frac=1, replace=False, random_state=512), df_ns.sample(frac=1, replace=False, random_state=256), 3336, 5429, "./data/sample_ann2_")    


## Dataset Analysis

import itertools
import collections

def get_data(path, lim=2):
    data = pd.read_csv(path, sep=',', error_bad_lines=False)
    res = {"uri":[], "text":[], "class":[]}
    for tag in data.columns[2:]:
        tmp = data[data[tag] >= lim]
        res["uri"] += tmp["uri"].tolist()
        res["text"] += tmp["text"].tolist()
        res["class"] += [tag]*len(tmp)
    res = pd.DataFrame(res).sort_values(by="uri")
    return res

# keep only "Sensibile" and "Non sensibile" as "sens" and "ns"
def get_two_classes(df):
    res = df[df["class"].isin(["Sensibile", "Non sensibile"])].copy()
    res.loc[res["class"] == "Sensibile", "class"] = "sens"
    res.loc[res["class"] == "Non sensibile", "class"] = "ns"
    return res

STOPWORDS = set(stopwords.words("english"))
STEMMER = SnowballStemmer('english')
LEMMATIZER = WordNetLemmatizer()

def clean_text(text):
    ## Remove puncuation
    text = text.translate(string.punctuation)
    text = word_tokenize(text)  # tokenize text
    ## Remove english stop words
    #stop_words = set(stopwords.words("english"))

    text = [w for w in text if w not in STOPWORDS and w.lower() not in STOPWORDS]  # remove stop words
    text = " ".join(text)

    # Clean the text
    text = re.sub("[^a-zA-z0-9\s]", " ", text)
    # text = re.sub("[^a-zA-Z#]", " ", text)
    text = word_tokenize(text)  # tokenize text
    stemmer = SnowballStemmer('english')
    stemmed_words = [stemmer.stem(word) for word in text]
    text = " ".join(stemmed_words)

    return text

#stem_or_lemma: 1 for stemming, 2 for lemmatization, other for none
def clean_text(text, remove_sw=False, stem_or_lemma=0):
    def remove_url(txt):
        return " ".join(re.sub("([^0-9A-Za-z \t])|(\w+:\/\/\S+)", "", txt).split())
    
    res = remove_url(text).lower().split()
    # Remove english stop words
    if remove_sw:
        #stop_words = set(stopwords.words("english"))
        res = [word for word in res if not word in STOPWORDS]
    if stem_or_lemma == 1: # stemming
        #stemmer = SnowballStemmer('english')
        res = [STEMMER.stem(word) for word in res]
        #res = " ".join([stemmer.stem(word) for word in res])
    elif stem_or_lemma == 2: # lemmatization
        #lemmatizer = WordNetLemmatizer()
        res = [LEMMATIZER.lemmatize(word, pos='v') for word in res]

    return res#.split() #remove_url(text).lower().split()

def get_df_stats(df):
    df_mean = df.groupby("word").mean().reset_index().sort_values(by="all", ascending=False).reset_index(drop=True)
    df_mean.columns = ["word"]+[x+"_mean" for x in df_mean.columns[1:]]
    df_std = df.groupby("word").std().fillna(0).reset_index()
    df_std.columns = ["word"]+[x+"_std" for x in df_std.columns[1:]]
    return df_mean.merge(df_std)

def get_df_stats_2(data, main_col):
    words_sens = list(itertools.chain(*data[data['class'] == 'sens'][main_col].values.tolist()))
    #words_ns = list(itertools.chain(*data[data['class'] == 'ns'][main_col].values.tolist()))
    words_all = list(itertools.chain(*data[main_col].values.tolist()))
    wordcount_sens = collections.Counter(words_sens)
    #wordcount_ns = collections.Counter(words_ns)
    wordcount_all = collections.Counter(words_all)
    print("Top words", wordcount_all.most_common(20),"\n")
    top_wc_dict = {"word": [], "all": [], "sens": [], "ns": [], "perc_sens": [], "perc_ns": []}
    for word in wordcount_all:
        num_all = wordcount_all[word]
        num_sens = wordcount_sens[word]
        perc_sens = round((num_sens*100)/num_all,2)
        #print(item[0], item[1], perc_sens, round((100.0-perc_sens), 2))
        top_wc_dict["word"].append(word)
        top_wc_dict["all"].append(num_all)
        top_wc_dict["sens"].append(num_sens)
        top_wc_dict["ns"].append(num_all-num_sens)
        top_wc_dict["perc_sens"].append(perc_sens)
        top_wc_dict["perc_ns"].append(round((100.0-perc_sens), 2))
    return pd.DataFrame(top_wc_dict)


### [Sens] Datasets

data = get_data("./data/annotation_results.csv", lim=2)
data = get_two_classes(data)
data['text_cleaned_stem'] = data['text'].map(lambda x: clean_text(x, remove_sw=True, stem_or_lemma=1))
data = get_df_stats_2(data, 'text_cleaned_stem').sort_values(by="all", ascending=False).reset_index(drop=True)
data

main_cols = ["word", "all", "perc_sens", "perc_ns"]
rename_cols = {"perc_sens":"%_sens", "perc_ns":"%_ns"}
top_values = 20
print("##  TOP", top_values, "all words")
print(data[main_cols][:top_values])
print("\n##  TOP", top_values, "sensible words")
main_cols = ["word", "sens", "perc_sens", "perc_ns"]
print(data.sort_values(by="sens", ascending=False)[main_cols].rename(columns=rename_cols)[:top_values])
print("\n##  TOP", top_values, "not sensible words")
main_cols = ["word", "ns", "perc_sens", "perc_ns"]
print(data.sort_values(by="ns", ascending=False)[main_cols].rename(columns=rename_cols)[:top_values])
print("\n##  TOP", top_values, "percentage sensible words")
main_cols = ["word", "all", "perc_sens", "perc_ns"]
print(data[data["all"]>=100].sort_values(by="perc_sens", ascending=False)[main_cols].rename(columns=rename_cols)[:top_values])
print("\n##  TOP", top_values, "percentage not sensible words")
main_cols = ["word", "all", "perc_sens", "perc_ns"]
print(data[data["all"]>=100].sort_values(by="perc_ns", ascending=False)[main_cols].rename(columns=rename_cols)[:top_values])


data = get_data("./data/annotation_results.csv", lim=3)
data = get_two_classes(data)
data['text_cleaned_stem'] = data['text'].map(lambda x: clean_text(x, remove_sw=True, stem_or_lemma=1))
data = get_df_stats_2(data, 'text_cleaned_stem').sort_values(by="all", ascending=False).reset_index(drop=True)
data

main_cols = ["word", "all", "perc_sens", "perc_ns"]
rename_cols = {"perc_sens":"%_sens", "perc_ns":"%_ns"}
top_values = 20
print("##  TOP", top_values, "all words")
print(data[main_cols][:top_values])
print("\n##  TOP", top_values, "sensible words")
main_cols = ["word", "sens", "perc_sens", "perc_ns"]
print(data.sort_values(by="sens", ascending=False)[main_cols].rename(columns=rename_cols)[:top_values])
print("\n##  TOP", top_values, "not sensible words")
main_cols = ["word", "ns", "perc_sens", "perc_ns"]
print(data.sort_values(by="ns", ascending=False)[main_cols].rename(columns=rename_cols)[:top_values])
print("\n##  TOP", top_values, "percentage sensible words")
main_cols = ["word", "all", "perc_sens", "perc_ns"]
print(data[data["all"]>=100].sort_values(by="perc_sens", ascending=False)[main_cols].rename(columns=rename_cols)[:top_values])
print("\n##  TOP", top_values, "percentage not sensible words")
main_cols = ["word", "all", "perc_sens", "perc_ns"]
print(data[data["all"]>=100].sort_values(by="perc_ns", ascending=False)[main_cols].rename(columns=rename_cols)[:top_values])


### [WH+TW] Datasets

data = pd.read_csv("./data/sample_ann2_01.csv")
data['text_cleaned_stem'] = data['text'].map(lambda x: clean_text(x, remove_sw=True, stem_or_lemma=1))
tmp = tmp_func(data, 'text_cleaned_stem')

data = pd.read_csv("./data/sample_ann2_02.csv")
data['text_cleaned_stem'] = data['text'].map(lambda x: clean_text(x, remove_sw=True, stem_or_lemma=1))
tmp = tmp_func(data, 'text_cleaned_stem').append(tmp)


data = pd.DataFrame()
for index in range(10):
    data_loc = pd.read_csv("./data/sample_ann2_"+(("0"+str(index+1))[-2:])+".csv")
    data_loc['text_cleaned_stem'] = data_loc['text'].map(lambda x: clean_text(x, remove_sw=True, stem_or_lemma=1))
    data = data.append(get_df_stats_2(data_loc, 'text_cleaned_stem'))

data = get_df_stats(data).sort_values(by="all_mean", ascending=False).reset_index(drop=True)
data


main_cols = ["word", "all_mean", "all_std", "perc_sens_mean", "perc_sens_std", "perc_ns_mean", "perc_ns_std"]
rename_cols = {"perc_sens_mean":"%_sens_mean", "perc_sens_std":"%_sens_std", "perc_ns_mean":"%_ns_mean", "perc_ns_std":"%_ns_std"}
top_values = 20
print("##  TOP", top_values, "all words")
print(data[main_cols].rename(columns=rename_cols)[:top_values])
print("\n##  TOP", top_values, "sensible words")
main_cols = ["word", "sens_mean", "sens_std", "perc_sens_mean", "perc_sens_std", "perc_ns_mean", "perc_ns_std"]
print(data.sort_values(by="sens_mean", ascending=False)[main_cols].rename(columns=rename_cols)[:top_values])
print("\n##  TOP", top_values, "not sensible words")
main_cols = ["word", "ns_mean", "ns_std", "perc_sens_mean", "perc_sens_std", "perc_ns_mean", "perc_ns_std"]
print(data.sort_values(by="ns_mean", ascending=False)[main_cols].rename(columns=rename_cols)[:top_values])
print("\n##  TOP", top_values, "percentage sensible words")
main_cols = ["word", "all_mean", "all_std", "perc_sens_mean", "perc_sens_std", "perc_ns_mean", "perc_ns_std"]
print(data[data["all_mean"]>=100].sort_values(by="perc_sens_mean", ascending=False)[main_cols].rename(columns=rename_cols)[:top_values])
print("\n##  TOP", top_values, "percentage not sensible words")
main_cols = ["word", "all_mean", "all_std", "perc_sens_mean", "perc_sens_std", "perc_ns_mean", "perc_ns_std"]
print(data[data["all_mean"]>=100].sort_values(by="perc_ns_mean", ascending=False)[main_cols].rename(columns=rename_cols)[:top_values])



## Create sets: training, validation and test

from lib.utils import get_data, get_two_classes, my_train_val_test_split

# training
train_ratio = 0.75
validation_ratio = 0.15
test_ratio = 0.10


for ann in [2,3]:
    print("## agreement on", ann, "annotators")
    # read data
    data = get_data("./data/annotation_results.csv", lim=ann)
    data = get_two_classes(data)
    # create train, validation and test sets
    x_train, x_val, x_test, y_train, y_val, y_test = my_train_val_test_split(data['text'].values, data['class'].values, validation_ratio, test_ratio, 512)
    print("Size of: training -> "+str(len(x_train))+" ("+str(int(100*len(x_train)/(len(data))))+"%), val -> "+str(len(x_val))+" ("+str(int(100*len(x_val)/(len(data))))+"%), test -> "+str(len(x_test))+" ("+str(int(100*len(x_test)/(len(data))))+"%)")
    # save sets
    pd.DataFrame({"text": x_train, "class": y_train}).to_csv("./data/annotation_results__ann"+str(ann)+"_training.csv", index=False)
    pd.DataFrame({"text": x_val, "class": y_val}).to_csv("./data/annotation_results__ann"+str(ann)+"_validation.csv", index=False)
    pd.DataFrame({"text": x_test, "class": y_test}).to_csv("./data/annotation_results__ann"+str(ann)+"_test.csv", index=False)

for index in range(10):
    print("## sample", (index+1))
    # read data
    pathname = "./data/sample_ann2_"+(("0"+str(index+1))[-2:])
    data = pd.read_csv(pathname+".csv")
    # create train, validation and test sets
    x_train, x_val, x_test, y_train, y_val, y_test = my_train_val_test_split(data['text'].values, data['class'].values, validation_ratio, test_ratio, 512)
    print("Size of: training -> "+str(len(x_train))+" ("+str(int(100*len(x_train)/(len(data))))+"%), val -> "+str(len(x_val))+" ("+str(int(100*len(x_val)/(len(data))))+"%), test -> "+str(len(x_test))+" ("+str(int(100*len(x_test)/(len(data))))+"%)")
    # save sets
    pd.DataFrame({"text": x_train, "class": y_train}).to_csv(pathname+"_training.csv", index=False)
    pd.DataFrame({"text": x_val, "class": y_val}).to_csv(pathname+"_validation.csv", index=False)
    pd.DataFrame({"text": x_test, "class": y_test}).to_csv(pathname+"_test.csv", index=False)
