==> 01 - Dict class.ipynb <==
# Import scikit-learn libraries
from sklearn.linear_model import LogisticRegression
from sklearn.svm import LinearSVC
from sklearn.neighbors import KNeighborsClassifier
from sklearn.ensemble import RandomForestClassifier

# import libraries for charting and manipulations with datasets
import matplotlib.pyplot as plt
import seaborn as sns
import pandas as pd
import numpy as np

# import self libraries
from lib.utils import get_data, get_two_classes, my_train_val_test_split
from sklearn import metrics


# directories
main_path = "./"
results_path = main_path+"experiments_03/"

# classes
Y_feat_names = ["ns", "sens"]

# models
models = [{"name": "LR", "model": LogisticRegression(max_iter=1000)},
          {"name": "lSVC", "model": LinearSVC(dual=False)},
          {"name": "KNN", "model": KNeighborsClassifier()},
          {"name": "RF", "model": RandomForestClassifier(n_estimators=100, random_state=0)},
         ]

# key columns
X_cols_dict = [{"name": "X", "cols": ['PrivTtl', 'Intimacy', 'Law', 'NegativePrivacy', 'NormsRequisites', 'OpenVisible', 'OutcomeState', 'PrivateSecret', 'Restriction', 'pronoun', 'ppron', 'i', 'we', 'you', 'shehe', 'they', 'ipron', 'article', 'prep', 'auxverb', 'adverb', 'conj', 'negate', 'verb', 'adj', 'compare', 'interrog', 'number', 'quant', 'affect', 'posemo', 'negemo', 'anx', 'anger', 'sad', 'social', 'family', 'friend', 'female', 'male', 'cogproc', 'insight', 'cause', 'discrep', 'tentat', 'certain', 'differ', 'percept', 'see', 'hear', 'feel', 'bio', 'body', 'health', 'sexual', 'ingest', 'drives', 'affiliation', 'achieve', 'power', 'reward', 'risk', 'focuspast', 'focuspresent', 'focusfuture', 'relativ', 'motion', 'space', 'time', 'work', 'leisure', 'home', 'money', 'relig', 'death', 'informal', 'swear', 'netspeak', 'assent', 'nonflu', 'filler']},
               {"name": "XNP", "cols": ['pronoun', 'ppron', 'i', 'we', 'you', 'shehe', 'they', 'ipron', 'article', 'prep', 'auxverb', 'adverb', 'conj', 'negate', 'verb', 'adj', 'compare', 'interrog', 'number', 'quant', 'affect', 'posemo', 'negemo', 'anx', 'anger', 'sad', 'social', 'family', 'friend', 'female', 'male', 'cogproc', 'insight', 'cause', 'discrep', 'tentat', 'certain', 'differ', 'percept', 'see', 'hear', 'feel', 'bio', 'body', 'health', 'sexual', 'ingest', 'drives', 'affiliation', 'achieve', 'power', 'reward', 'risk', 'focuspast', 'focuspresent', 'focusfuture', 'relativ', 'motion', 'space', 'time', 'work', 'leisure', 'home', 'money', 'relig', 'death', 'informal', 'swear', 'netspeak', 'assent', 'nonflu', 'filler']},
               {"name": "XP", "cols": ['PrivTtl', 'Intimacy', 'Law', 'NegativePrivacy', 'NormsRequisites', 'OpenVisible', 'OutcomeState', 'PrivateSecret', 'Restriction']}
              ]


# training
train_ratio = 0.75
validation_ratio = 0.15
test_ratio = 0.10

def get_indexes(main_path, datadict_path, ann, train_ratio, validation_ratio, test_ratio, random_state, text=True):
    data_loc = get_data(main_path, lim=ann)
    data_loc = get_two_classes(data_loc)
    x_train, x_val, x_test, y_train, y_val, y_test = my_train_val_test_split(data_loc['uri'].values, data_loc['class'].values, validation_ratio, test_ratio, random_state)
    if text:
        data_loc = pd.read_csv(main_path)
        data_loc = data_loc[["text"]].reset_index(drop=True)
        data_loc[["id"]] = data_loc.index
        data_loc = data_loc.merge(pd.read_csv(datadict_path, sep=";"))
    else:
        data_loc = pd.read_csv(datadict_path, sep=";")
    res_train = pd.DataFrame({"id": x_train}).merge(data_loc)
    res_test = pd.DataFrame({"id": x_test}).merge(data_loc)
    return [x_train, x_test]

def get_indexes_1(main_path, ann, train_ratio, validation_ratio, test_ratio, random_state):
    data_loc = get_data(main_path, lim=ann)
    data_loc = get_two_classes(data_loc)
    x_train, x_val, x_test, y_train, y_val, y_test = my_train_val_test_split(data_loc['uri'].values, data_loc['class'].values, validation_ratio, test_ratio, random_state)
    return [x_train, x_test]

def get_indexes_2(main_path, train_ratio, validation_ratio, test_ratio, random_state):
    data_loc = pd.read_csv(main_path)
    x_train, x_val, x_test, y_train, y_val, y_test = my_train_val_test_split(data_loc['uri'].values, data_loc['class'].values, validation_ratio, test_ratio, random_state)
    return [x_train, x_test]

def get_indexes(data_df, train_ratio, validation_ratio, test_ratio, random_state):
    x_train, x_val, x_test, y_train, y_val, y_test = my_train_val_test_split(data_df['uri'].values, data_df['class'].values, validation_ratio, test_ratio, random_state)
    return [x_train, x_test]

def get_metrics(y_test, y_pred, Y_feat_names):
    res = {}
    res["accuracy"] = metrics.accuracy_score(y_test, y_pred)
    tmp = metrics.precision_score(y_test, y_pred, average=None)
    for index, cls in enumerate(Y_feat_names):
        res["precision_"+cls] = tmp[index]
    tmp = metrics.recall_score(y_test, y_pred, average=None)
    for index, cls in enumerate(Y_feat_names):
        res["recall_"+cls] = tmp[index]
    tmp = metrics.f1_score(y_test, y_pred, average=None)
    for index, cls in enumerate(Y_feat_names):
        res["f1_"+cls] = tmp[index]
    res["f1-micro"] = metrics.f1_score(y_test, y_pred, average="micro")
    res["f1-macro"] = metrics.f1_score(y_test, y_pred, average="macro")
    return res

def calculate_metrics(data_df, datadict_path, model_dict, data_x_dict, train_ratio, validation_ratio, test_ratio, random_state):
    data = pd.read_csv(datadict_path, sep=';', error_bad_lines=False)
    dataY = data['class'].values
    dataX = data[data_x_dict["cols"]].values
    [train_index, test_index] = get_indexes(data_df, train_ratio, validation_ratio, test_ratio, random_state)
    X_train_k, X_test_k = dataX[train_index], dataX[test_index]
    Y_train_k, Y_test_k = dataY[train_index], dataY[test_index]
    model = model_dict["model"]
    model.fit(X_train_k, Y_train_k)
    model.predict(X_test_k)
    y_pred = model.predict(X_test_k)
    met_dict_loc = get_metrics(Y_test_k, y_pred, Y_feat_names)
    met_dict_loc["mod"] = model_dict["name"]
    met_dict_loc["x_col"] = data_x_dict["name"]
    return met_dict_loc


met_dict_loc = []
for ann in [2,3]:
    print("#### Agreement on", ann, "annotators")
    data_orig_path = main_path+'data/annotation_results.csv'
    data_df = get_two_classes(get_data(data_orig_path, lim=ann))
    datadict_path = main_path+"data/dictclass/datadict"+str(ann)+".csv"
    #model = DummyClassifier(strategy="most_frequent")
    for model_dict in models:
        print(" - model:", model_dict["name"])
        for X_cols in X_cols_dict: 
            met_dict_inn = calculate_metrics(data_df, datadict_path, model_dict, X_cols, train_ratio, validation_ratio, test_ratio, 512)
            met_dict_inn["data"] = "ann"+str(ann)
            met_dict_loc.append(met_dict_inn)

met_dict_df_1 = pd.DataFrame(met_dict_loc)
met_dict_df_1

met_dict_loc = []
for index in range(10):
    print("#### Sample", (index+1))
    data_df = pd.read_csv(main_path+"data/sample_ann2_"+(("0"+str(index+1))[-2:])+".csv")
    data_df["uri"] = data_df.index
    datadict_path = main_path+"data/dictclass/whtw"+str(index+1)+".csv"
    for model_dict in models:
        print(" - model:", model_dict["name"])
        for X_cols in X_cols_dict: 
            met_dict_inn = calculate_metrics(data_df, datadict_path, model_dict, X_cols, train_ratio, validation_ratio, test_ratio, 512)
            met_dict_inn["data"] = "sample_"+str(("0"+str(index+1))[-2:])
            met_dict_loc.append(met_dict_inn)

met_dict_df_2 = pd.DataFrame(met_dict_loc)
met_dict_df_2

met_dict_df_1.to_csv(results_path+"dictclass_agreements.csv", index=False)
met_dict_df_2.to_csv(results_path+"dictclass_samples.csv", index=False)
