==> 02 - Bag of words.ipynb <==
import pandas as pd
import time

import re
from nltk.corpus import stopwords
from nltk.stem import SnowballStemmer
from nltk.stem import WordNetLemmatizer
from nltk.tokenize import word_tokenize

#from tensorflow.keras.preprocessing.text import Tokenizer
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.svm import LinearSVC
from sklearn.neighbors import KNeighborsClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn import metrics

from lib.utils import get_data, get_two_classes


# directories
main_path = "./"
dataset_filename = main_path+"data/annotation_results__ann"
dataset_filename_2 = main_path+"data/annotation_results"
dataset_wt_filename = main_path+"data/sample_ann2_"
results_path = main_path+"experiments_03/"

# classes
Y_feat_names = ["ns", "sens"]

# models
models = [{"name": "LR", "model": LogisticRegression()},
          {"name": "lSVC", "model": LinearSVC()},
          #{"name": "KNN", "model": KNeighborsClassifier()},
          {"name": "RF", "model": RandomForestClassifier(n_estimators=100, random_state=0)},
         ]

STOPWORDS = set(stopwords.words("english"))
STEMMER = SnowballStemmer('english')
LEMMATIZER = WordNetLemmatizer()

#stem_or_lemma: 1 for stemming, 2 for lemmatization, other for none
def clean_text(text, remove_sw=False, stem_or_lemma=0):
    def remove_url(txt):
        return " ".join(re.sub("([^0-9A-Za-z \t])|(\w+:\/\/\S+)", "", txt).split())
    
    res = remove_url(text).lower().split()
    # Remove english stop words
    if remove_sw:
        #stop_words = set(stopwords.words("english"))
        res = [word for word in res if not word in STOPWORDS]
    if stem_or_lemma == 1: # stemming
        #stemmer = SnowballStemmer('english')
        res = [STEMMER.stem(word) for word in res]
        #res = " ".join([stemmer.stem(word) for word in res])
    elif stem_or_lemma == 2: # lemmatization
        #lemmatizer = WordNetLemmatizer()
        res = [LEMMATIZER.lemmatize(word, pos='v') for word in res]

    return res#.split() #remove_url(text).lower().split()

def get_metrics(y_test, y_pred, Y_feat_names):
    res = {}
    res["accuracy"] = metrics.accuracy_score(y_test, y_pred)
    tmp = metrics.precision_score(y_test, y_pred, average=None)
    for index, cls in enumerate(Y_feat_names):
        res["precision_"+cls] = tmp[index]
    res["precision-micro"] = metrics.precision_score(y_test, y_pred, average="micro")
    res["precision-macro"] = metrics.precision_score(y_test, y_pred, average="macro")
    tmp = metrics.recall_score(y_test, y_pred, average=None)
    for index, cls in enumerate(Y_feat_names):
        res["recall_"+cls] = tmp[index]
    tmp = metrics.f1_score(y_test, y_pred, average=None)
    for index, cls in enumerate(Y_feat_names):
        res["f1_"+cls] = tmp[index]
    res["f1-micro"] = metrics.f1_score(y_test, y_pred, average="micro")
    res["f1-macro"] = metrics.f1_score(y_test, y_pred, average="macro")
    return res

## [Sens] BoW: Execution and Performances
using as training sets Sens2 and Sens3, while as test sets Sens2, Sens3 and WH+TW

met_dict = []
experiment_time = time.perf_counter()
for ann in [2,3]:
    print("#### Agreement on", ann, "annotators")
    # get data
    train_loc = pd.read_csv(dataset_filename+str(ann)+"_training.csv")
    test_loc = pd.read_csv(dataset_filename+str(ann)+"_test.csv")
    # clean and stemmming
    train_loc['text_cleaned_stem'] = train_loc['text'].map(lambda x: " ".join(clean_text(x, remove_sw=True, stem_or_lemma=1)))
    test_loc['text_cleaned_stem'] = test_loc['text'].map(lambda x: " ".join(clean_text(x, remove_sw=True, stem_or_lemma=1)))
    # bag of words
    tfidfconverter = TfidfVectorizer()
    X_train = tfidfconverter.fit_transform(train_loc['text_cleaned_stem']).toarray()
    X_test = tfidfconverter.transform(test_loc['text_cleaned_stem']).toarray()
    # classes
    y_train = train_loc["class"].tolist()
    y_test = test_loc["class"].tolist()
    for model_dict in models:
        model_name = model_dict["name"]
        model = model_dict["model"]
        model_time = time.perf_counter()
        print(" - model:", model_name)
        # training
        model.fit(X_train, y_train)
        # prediction
        y_pred = model.predict(X_test)
        # model evaluation on test set
        met_dict_loc = get_metrics(y_test, y_pred, Y_feat_names)
        met_dict_loc["set"] = "test"
        met_dict_loc["data"] = "ann"+str(ann)
        met_dict_loc["mod"] = model_name
        met_dict.append(met_dict_loc)
        # model evaluation on other sets
        for index in range(10):
            test_loc_2 = pd.read_csv(dataset_wt_filename+(("0"+str(index+1))[-2:])+".csv")
            test_loc_2['text_cleaned_stem'] = test_loc_2['text'].map(lambda x: " ".join(clean_text(x, remove_sw=True, stem_or_lemma=1)))
            X_test_loc = tfidfconverter.transform(test_loc_2['text_cleaned_stem']).toarray()
            y_test_loc = test_loc_2["class"].tolist()
            y_pred_loc = model.predict(X_test_loc)
            met_dict_loc = get_metrics(y_test_loc, y_pred_loc, Y_feat_names)
            met_dict_loc["set"] = "sample_"+str(("0"+str(index+1))[-2:])
            met_dict_loc["data"] = "ann"+str(ann)
            met_dict_loc["mod"] = model_name
            met_dict.append(met_dict_loc)
        other_ann = 2
        if (ann == 2):
            other_ann = 3
        test_loc_2 = get_data(dataset_filename_2+".csv", lim=other_ann)
        test_loc_2 = get_two_classes(test_loc_2)
        test_loc_2['text_cleaned_stem'] = test_loc_2['text'].map(lambda x: " ".join(clean_text(x, remove_sw=True, stem_or_lemma=1)))
        X_test_loc = tfidfconverter.transform(test_loc_2['text_cleaned_stem']).toarray()
        y_test_loc = test_loc_2["class"].tolist()#pd.get_dummies(pd.DataFrame({"class": test_loc_2["class"].tolist()})["class"])[Y_feat_names].values
        y_pred_loc = model.predict(X_test_loc)
        met_dict_loc = get_metrics(y_test_loc, y_pred_loc, Y_feat_names)
        met_dict_loc["set"] = "ann"+str(other_ann)
        met_dict_loc["data"] = "ann"+str(ann)
        met_dict_loc["mod"] = model_name
        met_dict.append(met_dict_loc)
        model_time = time.perf_counter() - model_time
        print("   time:", time.strftime("/%d, %H:%M:%S",time.gmtime(model_time)))

experiment_time = time.perf_counter() - experiment_time
print("Experiment time in seconds:", int(experiment_time))
print("Experiment time:", time.strftime("/%d, %H:%M:%S",time.gmtime(experiment_time)))

met_dict_df_1 = pd.DataFrame(met_dict)
met_dict_df_1

met_dict_df_1["set"][0]

def get_df_stats_test(df, mod_name):
    df_loc = df[(df["set"] == "test") & (df["data"] == "ann2") & (df["mod"] == mod_name)][df.columns[:-3]]
    df_res = df_loc.T.reset_index()
    df_res.columns = ["metric", "ann2"]
    df_loc = df[(df["set"] == "test") & (df["data"] == "ann3") & (df["mod"] == mod_name)][df.columns[:-3]]
    df_loc = df_loc.T.reset_index()
    df_res["ann3"] = df_loc[df_loc.columns[-1]]
    return df_res

def get_df_stats_test_rev(df, mod_name):
    df_loc = df[(df["set"] == "ann3") & (df["data"] == "ann2") & (df["mod"] == mod_name)][df.columns[:-3]]
    df_res = df_loc.T.reset_index()
    df_res.columns = ["metric", "ann2 -> ann3"]
    df_loc = df[(df["set"] == "ann2") & (df["data"] == "ann3") & (df["mod"] == mod_name)][df.columns[:-3]]
    df_loc = df_loc.T.reset_index()
    df_res["ann3 -> ann2"] = df_loc[df_loc.columns[-1]]
    return df_res

def get_df_stats(df, ann, mod_name):
    df_loc = df[(df["set"].str[0] == 's') & (df["data"] == "ann"+str(ann)) & (df["mod"] == mod_name)][df.columns[:-3]]
    df_res = df_loc.mean().reset_index(drop=False)
    df_res[1] = df_loc.std().reset_index(drop=False)[0]
    df_res.columns = ["metric", "mean", "std"]
    return df_res

model_names = [x["name"] for x in models]

print("##  metrics on test set")
for mod_name in model_names:
    print("####  model:", mod_name)
    print(get_df_stats_test(met_dict_df_1, mod_name))

print("##  metrics on the other set")
for mod_name in model_names:
    print("####  model:", mod_name)
    print(get_df_stats_test_rev(met_dict_df_1, mod_name))
    
for ann in [2,3]:
    print("\n##  agreement on", ann, "annotators: test on 10 samples")
    for mod_name in model_names:
        print("####  model:", mod_name)
        print(get_df_stats(met_dict_df_1, ann, mod_name))

## [WH+TW] BoW: Execution and Performances
using as training sets Sens2 and Sens3, while as test sets Sens2, Sens3 and WH+TW

met_dict = []
experiment_time = time.perf_counter()
for index in range(10):
    # get data
    train_loc = pd.read_csv(dataset_wt_filename+(("0"+str(index+1))[-2:])+"_training.csv")
    test_loc = pd.read_csv(dataset_wt_filename+(("0"+str(index+1))[-2:])+"_test.csv")
    # clean and stemmming
    train_loc['text_cleaned_stem'] = train_loc['text'].map(lambda x: " ".join(clean_text(x, remove_sw=True, stem_or_lemma=1)))
    test_loc['text_cleaned_stem'] = test_loc['text'].map(lambda x: " ".join(clean_text(x, remove_sw=True, stem_or_lemma=1)))
    # bag of words
    tfidfconverter = TfidfVectorizer()
    X_train = tfidfconverter.fit_transform(train_loc['text_cleaned_stem']).toarray()
    X_test = tfidfconverter.transform(test_loc['text_cleaned_stem']).toarray()
    # classes
    y_train = train_loc["class"].tolist()
    y_test = test_loc["class"].tolist()
    for model_dict in models:
        model_name = model_dict["name"]
        model = model_dict["model"]
        model_time = time.perf_counter()
        print(" - model:", model_name)
        # training
        model.fit(X_train, y_train)
        # prediction
        y_pred = model.predict(X_test)
        # model evaluation on test set
        met_dict_loc = get_metrics(y_test, y_pred, Y_feat_names)
        met_dict_loc["set"] = "test"
        met_dict_loc["data"] = "sample_"+str(("0"+str(index+1))[-2:])
        met_dict_loc["mod"] = model_name
        met_dict.append(met_dict_loc)
        # model evaluation on other sets
        for ann in [2,3]:
            # get data
            test_loc_2 = get_data(dataset_filename_2+".csv", lim=ann)
            test_loc_2 = get_two_classes(test_loc_2)
            test_loc_2['text_cleaned_stem'] = test_loc_2['text'].map(lambda x: " ".join(clean_text(x, remove_sw=True, stem_or_lemma=1)))
            X_test_loc = tfidfconverter.transform(test_loc_2['text_cleaned_stem']).toarray()
            y_test_loc = test_loc_2["class"].tolist()
            y_pred_loc = model.predict(X_test_loc)
            met_dict_loc = get_metrics(y_test_loc, y_pred_loc, Y_feat_names)
            met_dict_loc["set"] = "ann"+str(ann)
            met_dict_loc["data"] = "sample_"+str(("0"+str(index+1))[-2:])
            met_dict_loc["mod"] = model_name
            met_dict.append(met_dict_loc)
            model_time = time.perf_counter() - model_time
        print("   time:", time.strftime("/%d, %H:%M:%S",time.gmtime(model_time)))

experiment_time = time.perf_counter() - experiment_time
print("Experiment time in seconds:", int(experiment_time))
print("Experiment time:", time.strftime("/%d, %H:%M:%S",time.gmtime(experiment_time)))

met_dict_df_2 = pd.DataFrame(met_dict)
met_dict_df_2


def get_df_stats_2(df):
    print("##  metrics on test set of 10 samples")
    for mod_name in model_names:
        print("####  model:", mod_name)
        #print(get_df_stats_test(met_dict_df_1, mod_name))
        df_loc = df[(df["set"] == "test") & (df["mod"] == mod_name)][df.columns[:-2]]
        df_res = df_loc.mean().reset_index(drop=False)
        df_res[1] = df_loc.std().reset_index(drop=False)[0]
        df_res.columns = ["metric", "mean", "std"]
        print(df_res)
        for ann in [2,3]:
            print("\n##  10 samples using as test set the agreement on", ann, "annotators")
            #print("####  model:", mod_name)
            df_loc = df[(df["set"] == "ann"+str(ann)) & (df["mod"] == mod_name)][df.columns[:-3]]
            df_res = df_loc.mean().reset_index(drop=False)
            df_res[1] = df_loc.std().reset_index(drop=False)[0]
            df_res.columns = ["metric", "mean", "std"]
            print(df_res)
    return

get_df_stats_2(met_dict_df_2)

met_dict_df_1.to_csv(results_path+"bow_agreements.csv", index=False)
met_dict_df_2.to_csv(results_path+"bow_samples.csv", index=False)
