==> 05 - Analysis of BERT.ipynb <==
import pandas as pd

## [Sens] Performances of BERT
using as training sets Sens2 and Sens3, while as test sets Sens2, Sens3 and WH+TW

df_res = pd.DataFrame()
for ann in [2,3]:
    df_res = df_res.append(pd.read_csv("./00 - Sensitag-BERT - our/ann"+str(ann)+".csv"))
df_res["precision-macro"] = (df_res["precision_ns"]+df_res["precision_sens"])/2
df_res = df_res[list(df_res.columns[:4])+[df_res.columns[-1]]+list(df_res.columns[4:-1])]
df_res


def get_df_stats_test(df):
    df_loc = df[(df["set"] == "test") & (df["ann"] == 2)][df.columns[:-2]]
    df_res = df_loc.T.reset_index()
    df_res.columns = ["metric", "ann2"]
    df_loc = df[(df["set"] == "test") & (df["ann"] == 3)][df.columns[:-2]]
    df_loc = df_loc.T.reset_index()
    df_res["ann3"] = df_loc[df_loc.columns[-1]]
    return df_res

def get_df_stats(df, ann):
    df_loc = df[(df["set"] != "test") & (df["ann"] == ann)][df.columns[:-2]]
    df_res = df_loc.mean().reset_index(drop=False)
    df_res[1] = df_loc.std().reset_index(drop=False)[0]
    df_res.columns = ["metric", "mean", "std"]
    return df_res

print("##  metrics on test set")
#for mod_name in model_names:
#    print("####  model:", mod_name)
print(get_df_stats_test(df_res))
#"""
for ann in [2,3]:
    print("\n##  agreement on", ann, "annotators: test on 10 samples")
    #for mod_name in model_names:
    #    print("####  model:", mod_name)
    print(get_df_stats(df_res, ann))
#"""


print("[sens2 -> sens3] precision macro:", (0.962588+0.910811)/2)
print("[sens3 -> sens2] precision macro:", (0.849354+0.78269)/2)

## [WH+TW] Performances of BERT
using as training sets WH+TW, while as test sets Sens2, Sens3 and WH+TW

df_res = pd.DataFrame()
for index in range(6):
    df_res = df_res.append(pd.read_csv("./01 - Sensitag-BERT - sample/sample_"+str(("0"+str(index+1))[-2:])+".csv"))
df_res["precision-macro"] = (df_res["precision_ns"]+df_res["precision_sens"])/2
df_res = df_res[list(df_res.columns[:4])+[df_res.columns[-1]]+list(df_res.columns[4:-1])]
df_res


def get_df_stats_2(df):
    print("##  metrics on test set of 10 samples")
    #for mod_name in model_names:
    #    print("####  model:", mod_name)
        #print(get_df_stats_test(met_dict_df_1, mod_name))
    df_loc = df[(df["set"] == "test")][df.columns[:-2]]
    df_res = df_loc.mean().reset_index(drop=False)
    df_res[1] = df_loc.std().reset_index(drop=False)[0]
    df_res.columns = ["metric", "mean", "std"]
    print(df_res)
    for ann in [2,3]:
        print("\n##  10 samples using as test set the agreement on", ann, "annotators")
        #print("####  model:", mod_name)
        df_loc = df[(df["set"] == "ann"+str(ann))][df.columns[:-2]]
        df_res = df_loc.mean().reset_index(drop=False)
        df_res[1] = df_loc.std().reset_index(drop=False)[0]
        df_res.columns = ["metric", "mean", "std"]
        print(df_res)
    return

get_df_stats_2(df_res)

## Gain over BoW

results_path_bow = "./experiments_03/"

met_dict_df_bow = pd.read_csv(results_path_bow+"bow_samples.csv")
met_dict_df_bow = met_dict_df_bow[(met_dict_df_bow["mod"] == "lSVC") & (met_dict_df_bow["set"] == "test")]
met_dict_df_bow["data"] = met_dict_df_bow["data"].apply(lambda x: x[0:6]+x[7:])
#met_dict_df_bow
df_res_2 = df_res[df_res["set"] == "test"][["f1-macro", "set", "data"]].merge(met_dict_df_bow[["f1-macro", "set", "data", "mod"]], left_on="data", right_on="data", suffixes=('', '_bow'))
#df_res_2
df_res_2["gain"] = (df_res_2["f1-macro"]/df_res_2["f1-macro_bow"] - 1.0)*100
#df_res_2
df_res_2[["set", "mod", "gain"]].groupby(["set", "mod"]).agg({"gain":['mean','std']}).reset_index()
